M2-6838: [Mobile] Complete test coverage for NotificationDaysExtractor
